+++ integreat_cms/cms/views/pages/page_revision_view.py
@@ -173,11 +173,15 @@ class PageRevisionView(TemplateView):
         revision.pk = None
         revision.version = current_revision.version + 1
 
-        if request.POST.get("submit_draft"):
+        if "submit_draft" in request.POST:
+            if not request.user.has_perm("cms.publish_page_object", page):
+                raise PermissionDenied(
+                    f"{request.user!r} does not have the permission to restore {revision!r} of {page!r} as draft"
+                )
             revision.status = status.DRAFT
-        elif request.POST.get("submit_review"):
+        elif "submit_review" in request.POST:
             revision.status = status.REVIEW
-        elif request.POST.get("submit_public"):
+        elif "submit_public" in request.POST:
             if not request.user.has_perm("cms.publish_page_object", page):
                 raise PermissionDenied(
                     f"{request.user!r} does not have the permission to restore the public {revision!r} of {page!r}"