+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 UNRELEASED
 ----------
 
+* [ [#1445](https://github.com/digitalfabrik/integreat-cms/issues/1445) ] Allow only users with publish permission to unpublish page
+* [ [#1497](https://github.com/digitalfabrik/integreat-cms/issues/1516) ] Set older versions to draft when saved as draft
+* [ [#1550](https://github.com/digitalfabrik/integreat-cms/issues/1550) ] Fix status change when restoring revisions
+* [ [#1509](https://github.com/digitalfabrik/integreat-cms/issues/1509) ] Support legacy sitemap URL patterns
+* [ [#742](https://github.com/digitalfabrik/integreat-cms/issues/742) ] Make bounding box configurable per region
+* [ [#742](https://github.com/digitalfabrik/integreat-cms/issues/742) ] Automatically fetch region bounding boxes from Nominatim API
+* [ [#1517](https://github.com/digitalfabrik/integreat-cms/issues/1517) ] Set all pages to draft when duplicating regions
+
 
 2022.6.1
 --------